# internal/argo-controller/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "SPX Argo Ctrl HTTP API",
        "title": "SPX Argo Ctrl",
        "contact": {
            "name": "API Support",
            "url": "https://superphenix.net",
            "email": "contact@rayshift.net"
        },
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
        "/v1/{orgId}/{projectId}/argo": {
            "get": {
                "security": [
                    {
                        "Bearer": [
                            ""
                        ]
                    }
                ],
                "description": "List every argo apps for a project",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1",
                    "Argo"
                ],
                "summary": "List Argo Apps",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "orgId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Applications",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/argo-controller_internal_v1_models_view.Application"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": [
                            ""
                        ]
                    }
                ],
                "description": "Create an Argo App",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1",
                    "Argo"
                ],
                "summary": "Create an Argo App",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "orgId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Argo App",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/argo-controller_internal_v1_argo_app.CreateAppInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/{orgId}/{projectId}/argo/localId/{localId}": {
            "get": {
                "security": [
                    {
                        "Bearer": [
                            ""
                        ]
                    }
                ],
                "description": "Get Argo App by Local ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1",
                    "Argo"
                ],
                "summary": "Get Argo App by Local ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "orgId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Local ID",
                        "name": "localId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Application",
                        "schema": {
                            "$ref": "#/definitions/argo-controller_internal_v1_models_view.Application"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/{orgId}/{projectId}/argo/{effectiveId}": {
            "get": {
                "security": [
                    {
                        "Bearer": [
                            ""
                        ]
                    }
                ],
                "description": "Get Argo App by Effective ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1",
                    "Argo"
                ],
                "summary": "Get Argo App by Effective ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "orgId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Effective ID",
                        "name": "effectiveId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Application",
                        "schema": {
                            "$ref": "#/definitions/argo-controller_internal_v1_models_view.Application"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": [
                            ""
                        ]
                    }
                ],
                "description": "Update Argo App",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1",
                    "Argo"
                ],
                "summary": "Update Argo App",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "orgId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Effective ID",
                        "name": "effectiveId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Argo App",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/argo-controller_internal_v1_argo_app.UpdateAppInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": [
                            ""
                        ]
                    }
                ],
                "description": "Delete Argo App by Effective ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1",
                    "Argo"
                ],
                "summary": "Delete Argo App",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "orgId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Effective ID",
                        "name": "effectiveId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/{orgId}/{projectId}/mark": {
            "get": {
                "security": [
                    {
                        "Bearer": [
                            ""
                        ]
                    }
                ],
                "description": "Mark all argo resources inside a project for deletion",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1",
                    "GC"
                ],
                "summary": "Mark for deletion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "orgId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "argo-controller_internal_v1_argo_app.AppSource": {
            "type": "object",
            "properties": {
                "chart": {
                    "type": "string"
                },
                "helm": {
                    "$ref": "#/definitions/v1alpha1.ApplicationSourceHelm"
                },
                "path": {
                    "type": "string"
                },
                "plugin": {
                    "$ref": "#/definitions/v1alpha1.ApplicationSourcePlugin"
                },
                "repoUrl": {
                    "type": "string"
                },
                "targetRevision": {
                    "type": "string"
                }
            }
        },
        "argo-controller_internal_v1_argo_app.CreateAppInfo": {
            "type": "object",
            "properties": {
                "general": {
                    "type": "object",
                    "properties": {
                        "appName": {
                            "type": "string"
                        },
                        "destination": {
                            "type": "string"
                        }
                    }
                },
                "orgID": {
                    "type": "string"
                },
                "projectID": {
                    "type": "string"
                },
                "resourceEffectiveID": {
                    "type": "string"
                },
                "resourceLocalID": {
                    "type": "string"
                },
                "spec": {
                    "type": "object",
                    "properties": {
                        "ignoreDifferences": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/v1alpha1.ResourceIgnoreDifferences"
                            }
                        },
                        "source": {
                            "$ref": "#/definitions/argo-controller_internal_v1_argo_app.AppSource"
                        }
                    }
                }
            }
        },
        "argo-controller_internal_v1_argo_app.UpdateAppInfo": {
            "type": "object",
            "properties": {
                "ignoreDifferences": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1alpha1.ResourceIgnoreDifferences"
                    }
                },
                "source": {
                    "$ref": "#/definitions/argo-controller_internal_v1_argo_app.AppSource"
                }
            }
        },
        "argo-controller_internal_v1_models_view.AppView": {
            "type": "object",
            "properties": {
                "apiVersion": {
                    "description": "APIVersion defines the versioned schema of this representation of an object.\nServers should convert recognized schemas to the latest internal value, and\nmay reject unrecognized values.\nMore info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources\n+optional",
                    "type": "string"
                },
                "kind": {
                    "description": "Kind is a string value representing the REST resource this object represents.\nServers may infer this from the endpoint the client submits requests to.\nCannot be updated.\nIn CamelCase.\nMore info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds\n+optional",
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/argo-controller_internal_v1_models_view.ObjectMeta"
                },
                "operation": {
                    "$ref": "#/definitions/v1.Operation"
                },
                "spec": {
                    "$ref": "#/definitions/argo-controller_internal_v1_models_view.ApplicationSpec"
                },
                "status": {
                    "$ref": "#/definitions/argo-controller_internal_v1_models_view.ApplicationStatus"
                }
            }
        },
        "argo-controller_internal_v1_models_view.Application": {
            "type": "object",
            "properties": {
                "app": {
                    "$ref": "#/definitions/argo-controller_internal_v1_models_view.AppView"
                },
                "codeAZ": {
                    "type": "string"
                },
                "eid": {
                    "description": "effective ID",
                    "type": "string"
                },
                "gitops": {
                    "type": "string"
                },
                "id": {
                    "description": "local ID",
                    "type": "string"
                },
                "resourceName": {
                    "description": "human-readable name",
                    "type": "string"
                },
                "resourceTypeId": {
                    "type": "string"
                }
            }
        },
...
        "argo-controller_internal_v1_models_view.ApplicationSpec": {
            "type": "object",
            "properties": {
                "source": {
                    "description": "Source is a reference to the location of the application's manifests or chart",
                    "allOf": [
                        {
                            "$ref": "#/definitions/argo-controller_internal_v1_models_view.ApplicationSource"
                        }
                    ]
                }
            }
        },
        "argo-controller_internal_v1_models_view.ApplicationStatus": {
            "type": "object",
            "properties": {
                "conditions": {
                    "description": "Conditions is a list of currently observed application conditions",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1alpha1.ApplicationCondition"
                    }
                },
                "health": {
                    "description": "Health contains information about the application's current health status",
                    "allOf": [
                        {
                            "$ref": "#/definitions/v1alpha1.AppHealthStatus"
                        }
                    ]
                },
                "resources": {
                    "description": "Resources is a list of Kubernetes resources managed by this application",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.ResourceStatus"
                    }
                },
                "sync": {
                    "description": "Sync contains information about the application's current sync status",
                    "allOf": [
                        {
                            "$ref": "#/definitions/v1alpha1.SyncStatus"
                        }
                    ]
                }
            }
        },
        "argo-controller_internal_v1_models_view.ObjectMeta": {
            "type": "object",
            "properties": {
                "creationTimestamp": {
                    "type": "string"
                },
                "generateName": {
                    "type": "string"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "ownerReferences": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.OwnerReference"
                    }
                }
            }
        },
...
        "v1.Operation": {
            "type": "string",
            "enum": [
                "CREATE",
                "UPDATE",
                "DELETE",
                "CONNECT"
            ],
            "x-enum-varnames": [
                "Create",
                "Update",
                "Delete",
                "Connect"
            ]
        },
...
        "v1alpha1.ResourceIgnoreDifferences": {
            "type": "object",
            "properties": {
                "group": {
                    "type": "string"
                },
                "jqPathExpressions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "jsonPointers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "kind": {
                    "type": "string"
                },
                "managedFieldsManagers": {
                    "description": "ManagedFieldsManagers is a list of trusted managers. Fields mutated by those managers will take precedence over the\ndesired state defined in the SCM and won't be displayed in diffs",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Auth Token to have access to Argo Ctrl. \"Bearer [Auth Token]\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "scopes": {
                "organizationbillingread": "Grants read access to organization billing",
                "organizationbillingwrite": "Grants write access to organization billing",
                "organizationiamread": "Grants read access to organization IAM",
                "organizationiamwrite": "Grants write access to organization IAM",
                "organizationprojectmanagement": "Grants project management access in organization",
                "organizationread": "Grants read access to organization",
                "organizationwrite": "Grants write access to organization",
                "projectdiskread": "Grants read access to Disk in project",
                "projectdiskwrite": "Grants write access to Disk in project",
                "projecteipread": "Grants read access to EIP in project",
                "projecteipwrite": "Grants write access to EIP in project",
                "projectinstancecontrol": "Grants control (start, stop) access to instance in project",
                "projectinstanceread": "Grants read access to instance in project",
                "projectinstanceterminal": "Grants terminal access to instance in project",
                "projectinstancewrite": "Grants write access to instance in project",
                "projectlbread": "Grants read access to LoadBalancer in project",
                "projectlbwrite": "Grants write access to LoadBalancer in project",
                "projectsnapshotread": "Grants read access to Snapshot and Instance Snapshot in project",
                "projectsnapshotwrite": "Grants write access to Snapshot and Instance Snapshot in project",
                "projectsshread": "Grants read access to SSH in project",
                "projectsshwrite": "Grants write access to SSH in project",
                "projectsubnetread": "Grants read access to Subnet in project",
                "projectsubnetwrite": "Grants write access to Subnet in project",
                "projectvpcread": "Grants read access to VPC in project",
                "projectvpcwrite": "Grants write access to VPC in project"
            }
        }
    }
}